% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{fits_dist_matrix}
\alias{fits_dist_matrix}
\title{Given a list of grouped fits, compute a distance matrix}
\usage{
fits_dist_matrix(group_fits)
}
\arguments{
\item{group_fits}{A list of grouped fits returned by {group_fit_series}}
}
\value{
A dist matrix of pairwise distances between each group-fit
}
\description{
Given a list of grouped fits, compute a distance matrix
}
