% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitSeries}
\alias{fitSeries}
\title{For backward compactability}
\usage{
fitSeries(history, parallel = T, kernel.type = "EXPN", .init_no = NA,
  .additional.PLN.final.fit = F, .fit_engine = "AMPL",
  single_event_cascade = F, lowerBound = NULL, upperBound = NULL,
  marked = T, ...)
}
\description{
For backward compactability
}
