% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{generate_features}
\alias{generate_features}
\title{Given a list of group-fits produced by 'group_fit_series', this function generates features
for each group-fit by summarizing the fitted parameters.}
\usage{
generate_features(list_fits)
}
\arguments{
\item{list_fits}{A list of group fits returned by {group_fit_series}}
}
\value{
A data frame of features for each group. If features are all -1, it means all the
fits of the group are NAs
}
\description{
Given a list of group-fits produced by 'group_fit_series', this function generates features
for each group-fit by summarizing the fitted parameters.
}
