% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_user_magnitude}
\alias{generate_user_magnitude}
\title{Function for sampling from the powerlaw distribution of user influence
it is the equivalent of the Richter-Gutenberg distribution in the Helmstetter model
the powerlaw distribution was determined from the twitter data, from the #retweets
alpha = 2.016, xmin = 1. Draw n values}
\usage{
generate_user_magnitude(n, alpha = 2.016, mmin = 1)
}
\arguments{
\item{n}{The number of samples to be generated}

\item{alpha}{Powerlaw distribution parameters}

\item{mmin}{Powerlaw distribution parameters}
}
\value{
A single number, a random user magnitude
}
\description{
Function for sampling from the powerlaw distribution of user influence
it is the equivalent of the Richter-Gutenberg distribution in the Helmstetter model
the powerlaw distribution was determined from the twitter data, from the #retweets
alpha = 2.016, xmin = 1. Draw n values
}
\examples{
generate_user_magnitude(n = 1)
}
