% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_a1}
\alias{get_a1}
\title{Calculating the expected size of first level of descendants}
\usage{
get_a1(model)
}
\arguments{
\item{model}{A model object provides data, model_type, observation_time
and model parameters}
}
\value{
a vector of the expected sizes of first level of descendants of the
given cascades
}
\description{
Calculating the expected size of first level of descendants
}
