% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_viral_score}
\alias{get_viral_score}
\title{Viral score is the total reaction of the system to a single promotion,
i.e. the expected cascade size started by a single event of magnitude}
\usage{
get_viral_score(model, mu)
}
\arguments{
\item{model}{A model object for computing the branching factor.}

\item{mu}{The magnitude of the initial event}
}
\value{
A single number, the viral score of the given model
}
\description{
Viral score is the total reaction of the system to a single promotion,
i.e. the expected cascade size started by a single event of magnitude
}
