% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{group_fit_series}
\alias{group_fit_series}
\title{Given a list of cascades, this function fits each cascade individually by calling {fit_series}.
If the given cascades are in a named list, the names will be regarded as groups and the result will be reformatted as a list of
group fits.}
\usage{
group_fit_series(
  data,
  model_type,
  observation_times = NULL,
  cores = 1,
  verbose = F,
  ...
)
}
\arguments{
\item{data}{a (named) list of data.frame(s) where each data.frame is an event cascade with
event tims and event magnitudes (optional). The list names (if present) will be used for grouping cascades with same names.}

\item{model_type}{a string representing the model type, e.g. EXP for Hawkes processes with an exponential kernel function}

\item{observation_times}{a list or an vector of observation times for the cascades in data}

\item{cores}{the number of cores used for parallel fitting, defaults to 1 (non-parallel)}

\item{...}{check the available arguments of {fit_series}}
}
\description{
Given a list of cascades, this function fits each cascade individually by calling {fit_series}.
If the given cascades are in a named list, the names will be regarded as groups and the result will be reformatted as a list of
group fits.
}
