% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{melt_snowflake}
\alias{melt_snowflake}
\title{Inversely transform a Twitter id back to its components (timestamp, datacentre id,
worker id and sequence id).}
\usage{
melt_snowflake(snowflake_id, twepoch = bit64::as.integer64("1288834974657"))
}
\arguments{
\item{snowflake_id}{a list of Twitter ids}

\item{twepoch}{time epoch. Defaults to 1288834974657}
}
\value{
A list of timestamp, datacentre id, worker id and sequence id
}
\description{
Inversely transform a Twitter id back to its components (timestamp, datacentre id,
worker id and sequence id).
}
