% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{new_hawkes}
\alias{new_hawkes}
\title{Create a new hawkes model with parameters available}
\usage{
new_hawkes(model_type, par = NULL, data = NULL, init_par = NULL,
  observation_time = NULL, lower_bound = NULL, upper_bound = NULL,
  model_vars = NULL)
}
\arguments{
\item{model_type}{A string indicates the model tyep, e.g. EXP for a Hawkes process
with an exponential kernel}

\item{par}{A named vector denotes the model parameters where the names are model
parameters and the values are the corresponding parameter values}

\item{data}{a list of data.frame(s) where each data.frame is an event cascade with event
tims and event magnitudes (optional)}

\item{init_par}{Initial parameter values used in fitting}

\item{observation_time}{the event cascades observation time. It is assumed that all cascades in data
are observed until a common time.}

\item{lower_bound}{model parameter lower bounds. A named vector where names are model parameters and
values are the lowest possible values.}

\item{model_vars}{a named list of extra variables provided to hawkes objects}

\item{uppper_bound}{model parameter upper bounds. A named vector where names are model parameters and
values are the largest possible values.}
}
\description{
Create a new hawkes model with parameters available
}
