% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_series}
\alias{plot_event_series}
\title{Plot a Hawkes process and its intensity function}
\usage{
plot_event_series(model, cascade_index = 1)
}
\arguments{
\item{model}{a model object where data, model_type and par are required}

\item{cascade_index}{determine which cascade in the list of cascades to plot
default to the first cascade}
}
\description{
Plot a Hawkes process and its intensity function
}
