% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kernel_function}
\alias{plot_kernel_function}
\title{Plot the kernel functions of Hawkes processes}
\usage{
plot_kernel_function(fitted_models)
}
\arguments{
\item{fitted_models}{A list of fitted model objects to plot the kernel functions}
}
\value{
A ggplot object
}
\description{
Plot the kernel functions of Hawkes processes
}
