% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{prepare_tmp_file}
\alias{prepare_tmp_file}
\title{Prepare the temporary auxilixry files for AMPL}
\usage{
prepare_tmp_file(type)
}
\arguments{
\item{type}{is one of "mod" (AMPL model file), "dat" (AMPL data file),
"run" (AMPL run file) and "res" (file hosts AMPL runned output)}
}
\description{
Prepare the temporary auxilixry files for AMPL
}
