% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{set_tmp_folder}
\alias{set_tmp_folder}
\title{Set up the folder for placing temporary files, defaults to /tmp}
\usage{
set_tmp_folder(path)
}
\arguments{
\item{path}{A string of path to the folder where you would like to place temporary files}
}
\description{
Set up the folder for placing temporary files, defaults to /tmp
}
