% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{setup_ampl}
\alias{setup_ampl}
\title{Set up the AMPL execution path.}
\usage{
setup_ampl(ampl_path)
}
\arguments{
\item{ampl_path}{a string of path to where the executable ampl binary is located. Please make sure ipopt binary is in the same path.}
}
\description{
Set up the AMPL execution path.
}
