% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_Hawkes_event_series}
\alias{generate_Hawkes_event_series}
\title{Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)}
\usage{
generate_Hawkes_event_series(params, model_type, alpha = 2.016,
  mmin = 1, M = 10000, Tmax = 10, filename = NULL,
  history_init = NULL, maxEvents = NA, immigrants = NULL)
}
\arguments{
\item{params}{a named vector of model parameters, K, alpha, beta, mmin, c, theta - parameters of the Hawkes kernel}

\item{M}{- magnitude of the initial event (in case no initial history
provided)}

\item{Tmax}{- maximum time of simulation.}

\item{filename}{- file to which save the CSV file with the simulation.}

\item{history_init}{An initial history can be provided (R structure obtained
from a previous call of this function). Its purpose is to allow custom
initializations and to continue simulation of stopped processes.}
}
\description{
Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)
}
