% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_hawkes_event_series}
\alias{generate_hawkes_event_series}
\title{Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)}
\usage{
generate_hawkes_event_series(
  model,
  par,
  model_type,
  sim_no = 1,
  cores = 1,
  Tmax = Inf,
  maxEvents = NULL,
  M = NULL,
  tol = 1e-05
)
}
\arguments{
\item{model}{A model class object with par and model_type presented. par and
model_type are not requird once this is given}

\item{par}{A named vector of model parameters, K, alpha, beta, mmin, c, theta - parameters of the Hawkes kernel}

\item{model_type}{Model type}

\item{sim_no}{The number of simulated cascades}

\item{cores}{The number of cores (processes) used for simulation}

\item{Tmax}{Maximum time of simulation.}

\item{maxEvents}{Maximum number of events to be simulated.}

\item{M}{Magnitude of the initial event}

\item{tol}{Simulation stops when intensity smaller than tol.}
}
\value{
A list of data.frames where each data.frame is a simualted event cascade with the given model
}
\description{
Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)
}
\examples{
generate_hawkes_event_series(model_type = 'EXP',
                             par = c(K = 0.9, theta = 1),
                             sim_no = 10, Tmax = Inf)
}
