% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_hawkes_event_series}
\alias{generate_hawkes_event_series}
\title{Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)}
\usage{
generate_hawkes_event_series(model, par, model_type, sim_no = 1,
  cores = 1, Tmax = Inf, maxEvents = NULL, M = NULL,
  history_init = NULL, tol = 1e-05)
}
\arguments{
\item{model}{a model class object with par and model_type presented. par and
model_type are not requird once this is given}

\item{par}{a named vector of model parameters, K, alpha, beta, mmin, c, theta - parameters of the Hawkes kernel}

\item{model_type}{model type}

\item{sim_no}{the number of simulated cascades}

\item{cores}{the number of cores (processes) used for simulation}

\item{Tmax}{maximum time of simulation.}

\item{maxEvents}{maximum number of events to be simulated.}

\item{M}{magnitude of the initial event (in case no initial history
provided)}

\item{history_init}{An initial history can be provided (R structure obtained
from a previous call of this function). Its purpose is to allow custom
initializations and to continue simulation of stopped processes.}

\item{tol}{simulation stops when intensity smaller than tol.}
}
\description{
Main function to generate a Hawkes process sequence. It allows intermediary
saves and continuing a stopped simulation. Creates a CSV file with two
columns, each row is an event: (magnitude, time)
}
