% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_branching_factor}
\alias{get_branching_factor}
\title{Branching factor is the expected number of events generated
by a single event.}
\usage{
get_branching_factor(model)
}
\arguments{
\item{model}{a model object for computing the branching factor.}
}
\description{
Branching factor is the expected number of events generated
by a single event.
}
