% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_hawkes_neg_likelihood_value}
\alias{get_hawkes_neg_likelihood_value}
\title{Compute the negative log-likelihood values of a given model on a list of given
event cascades.}
\usage{
get_hawkes_neg_likelihood_value(model, ...)
}
\arguments{
\item{model}{An object of a specific model class where the `data` and the `par` fields
are required}

\item{...}{further arguments passed to ampl}
}
\description{
Compute the negative log-likelihood values of a given model on a list of given
event cascades.
}
