% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_hawkes_neg_likelihood_value}
\alias{get_hawkes_neg_likelihood_value}
\title{Compute the negative log-likelihood values of a given model on a list of given
event cascades.}
\usage{
get_hawkes_neg_likelihood_value(
  model,
  ...,
  par,
  data,
  model_type,
  observation_time
)
}
\arguments{
\item{model}{An object of a specific model class where the `data` and the `par` fields
are required.}

\item{...}{Further arguments passed to ampl}

\item{par}{Hawkes model parameters}

\item{data}{A list of data.frames of event cascades}

\item{model_type}{The Hawkes model type}

\item{observation_time}{The observation time of the given event cascades}
}
\value{
A single number, the negative log-likelihood of the given model on data
}
\description{
Compute the negative log-likelihood values of a given model on a list of given
event cascades.
}
\examples{
data <- generate_hawkes_event_series(model_type = 'EXP',
                                     par = c(K = 0.9, theta = 1),
                                     sim_no = 10, Tmax = Inf)
fitted <- fit_series(data, 'EXP', observation_time = Inf)
data_test <- generate_hawkes_event_series(model_type = 'EXP',
                                          par = c(K = 0.9, theta = 1),
                                          sim_no = 10, Tmax = Inf)
get_hawkes_neg_likelihood_value(fitted, data = data_test)
}
