% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_intensity_at}
\alias{get_model_intensity_at}
\title{Compute the intensity value of a given model at time t}
\usage{
get_model_intensity_at(model, t, cascade_index = 1)
}
\arguments{
\item{model}{a model object for computing the intensity value}

\item{t}{the given time to compute the intensity}

\item{cascade_index}{determine which cascade in the list of cascades to compute, defaults to 1}
}
\description{
Compute the intensity value of a given model at time t
}
