% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_viral_score}
\alias{get_viral_score}
\title{Viral score is the total reaction of the system to a single promotion,
i.e. the expected cascade size started by a single event of magnitude $\mu$}
\usage{
get_viral_score(model)
}
\arguments{
\item{model}{a model object for computing the branching factor.}

\item{mu}{the magnitude of the initial event}
}
\description{
Viral score is the total reaction of the system to a single promotion,
i.e. the expected cascade size started by a single event of magnitude $\mu$
}
