% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{parse_raw_tweets_to_cascades}
\alias{parse_raw_tweets_to_cascades}
\title{This function extracts cascades from a given jsonl file where each line is a tweet
json object. Please refer to the Twitter developer documentation: https://developer.twitter.com/en/docs/tweets/data-dictionary/overview/tweet-object}
\usage{
parse_raw_tweets_to_cascades(path)
}
\arguments{
\item{path}{file path to the tweets jsonl file}
}
\description{
This function extracts cascades from a given jsonl file where each line is a tweet
json object. Please refer to the Twitter developer documentation: https://developer.twitter.com/en/docs/tweets/data-dictionary/overview/tweet-object
}
