% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_series}
\alias{plot_event_series}
\title{Plot a Hawkes process and its intensity function}
\usage{
plot_event_series(model, cascade = NULL, cascade_index = 1)
}
\arguments{
\item{model}{A model object where data, model_type and par are required}

\item{cascade}{This cascade data.frame will be used if provided}

\item{cascade_index}{Determine which cascade in the list of cascades to plot
default to the first cascade}
}
\value{
A ggplot object
}
\description{
Plot a Hawkes process and its intensity function
}
\examples{
par <- c(K = 0.95, theta = 1)
data <- generate_series(model_type = 'EXP',
                                     par = par,
                                     sim_no = 1, Tmax = Inf)
plot_event_series(new_hawkes(model_type = 'EXP',
                             par = par,
                             data = data))
}
