% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict_final_popularity}
\alias{predict_final_popularity}
\title{Predict the final popularity (event count) of give histories and
its model parameters.}
\usage{
predict_final_popularity(model)
}
\arguments{
\item{model}{a model object provides data, model_type, observation_time
and model parameters}
}
\value{
a vector of predicted final popularities whose length is the same
as the number of cascades in the provided model object
}
\description{
Predict the final popularity (event count) of give histories and
its model parameters.
}
