% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{setup_ampl}
\alias{setup_ampl}
\title{Set up the AMPL environment by downloading an AMPL demo version and the compiled
ipopt binary. Only supports UNIX compatible OSs.}
\usage{
setup_ampl(ampl_path)
}
\arguments{
\item{ampl_path}{The path where the AMPL folder will be placed}
}
\description{
Set up the AMPL environment by downloading an AMPL demo version and the compiled
ipopt binary. Only supports UNIX compatible OSs.
}
